backend/pyproject.toml
[tool.poetry]
name = "memecry"
version = "0.1.0"
description = ""
authors = ["crpier <crpier42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
python-multipart = "^0.0.5"
aiofiles = "^22.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlmodel = "^0.0.8"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.4"}
jinja2 = "^3.1.2"
babel = "^2.11.0"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.254"
black = "^22.12.0"

[tool.ruff]
extend-select = ["F", "E", "W", "C90", "I", "N", "UP", "YTT", "ANN", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "RUF"]
ignore = ["B008", "UP035", "S101", "A003", "ANN101"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
